# specs/002-fastapi-dashboard-api/contracts/openapi-spec.yaml
openapi: 3.0.3
info:
  title: RiseTrader Dashboard API
  description: |
    FastAPI service serving market data, forecasts, account information, and strategy allocations
    from PostgreSQL database to React dashboard with real-time MT4 integration.

    ## Real-Time Updates
    The API provides Server-Sent Events (SSE) endpoints for real-time data streaming:
    - `/api/stream/market-data` - Price updates
    - `/api/stream/account` - Account balance and positions
    - `/api/stream/forecasts` - New forecast availability

    ## Authentication
    ⚠️ Currently NO authentication implemented. API key + JWT authentication will be added in Phase 8-9.

    ## Rate Limiting
    Rate limiting is enforced via slowapi middleware. Default limits apply per IP address.

    ## Pagination
    List endpoints use keyset (cursor-based) pagination for consistent performance:
    - `page`: Page number (1-indexed)
    - `page_size`: Items per page (default: 50, max: 500)
    - `cursor`: Opaque cursor for next page (returned in `next_cursor` field)
  version: 1.0.0
  contact:
    name: RiseTrader Team
    email: support@risetrader.com
  license:
    name: Proprietary
servers:
  - url: http://localhost:8003/api
    description: Local development server
  - url: https://api.risetrader.com/api
    description: Production server (future)
...
paths:
...
  /forecasts/latest:
    get:
      tags:
        - forecasts
      summary: Get latest forecasts
      description: Retrieve the latest price forecasts for all symbols
      operationId: getLatestForecasts
      parameters:
        - name: symbol
          in: query
          required: false
          schema:
            type: string
        - name: horizon
          in: query
          required: false
          description: Forecast horizon filter
          schema:
            type: string
            enum: ["1h", "4h", "24h"]
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ForecastListResponse'
  /forecasts/{symbol}:
    get:
      tags:
        - forecasts
      summary: Get forecasts for specific symbol
      description: Retrieve all forecasts for a specific symbol
      operationId: getSymbolForecasts
      parameters:
        - name: symbol
          in: path
          required: true
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: page_size
          in: query
          schema:
            type: integer
            default: 50
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ForecastListResponse'

  # ==========================================
  # STRATEGIES ENDPOINTS
  # ==========================================
...
components:
  schemas:
...
    ForecastResponse:
      type: object
      properties:
        id:
          type: integer
        model_type:
          type: string
          example: "XGBoost"
        symbol:
          type: string
        forecast_horizon:
          type: string
          enum: ["1h", "4h", "24h"]
        predicted_price:
          type: number
          format: double
        confidence_level:
          type: number
          format: double
          example: 0.95
        prediction_timestamp:
          type: string
          format: date-time
        target_timestamp:
          type: string
          format: date-time
        actual_price:
          type: number
          format: double
          nullable: true
        prediction_error:
          type: number
          format: double
          nullable: true
    ForecastListResponse:
      type: object
      properties:
        forecasts:
          type: array
          items:
            $ref: '#/components/schemas/ForecastResponse'
        total:
          type: integer

    # ==========================================
    # STRATEGY SCHEMAS
    # ==========================================
...